Restructure dynamic backend files
